requirements.txt
tiktoken>=0.7.0
openai>=1.30.0
chromadb>=0.5.0

weaviate-client>=4.6.0

rank-bm25>=0.2.2
sentence-transformers>=3.0.0   
python-dotenv>=1.0.0
pydantic>=2.7.0
pyyaml>=6.0.1
rich>=13.7.0                    

ragas>=0.1.13                   
pytest>=8.2.0
